balance of the tree

+++ 14-binary_tree_balance.c
@@ -16,8 +16,8 @@ int binary_tree_balance(const binary_tree_t *tree)
 	if (!tree)
 		return (0);
 
-	left_height = tree->left ? binary_tree_height(tree->left) : -1;
-	right_height = tree->right ? binary_tree_height(tree->right) : -1;
+	left_height = tree->left ? (int)binary_tree_height(tree->left) : -1;
+	right_height = tree->right ? (int)binary_tree_height(tree->right) : -1;
 
 	return (left_height - right_height);
 }
